merge

+++ train_flux_simple.py
@@ -799,7 +799,7 @@ def main():
                 tensors_flat_t0 = hyper.flatten_cached_from_cache()
 
                 #Compute LoRAs at t=1, 2, 3, ... B
-                t_ = (torch.arange(B, device=accelerator.device) % B) + 1
+                t_ = (torch.arange(B, device=accelerator.device, dtype=torch.bfloat16) % B) + 1
                 hyper.compute_and_cache_loras(batch_prompts, t_)
                 tensors_flat_t1 = hyper.flatten_cached_from_cache()
